data/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
qdrant-client==1.7.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
sentence-transformers>=2.3.0,<3.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
torch>=2.2.0,<2.3.0
numpy==1.24.3

